Python-Flask-MNIST-sample-app/app/requirements.txt
Flask==1.0.2
watson-machine-learning-client==1.0.346
Pillow==5.3.0
werkzeug==0.14.1
